braids and knots part IV - float64 fixes and README
## float32 → float64 on the 14-gon vertex table and IDENTITY. 
These sat directly above trot_memo, which was already float64, so the file was mixing precisions.

It turned out to be a real bug rather than tidying. braid_words.py, braid_words_bn.py and braided_tiling.py identify shared vertices with np.round(p, 6) as a dict key. Coordinates grow like λⁿ while float32's relative precision is fixed, so at depth the keys stop matching:
```
iterations	float32 shared / boundary	float64 shared / boundary
1	40 / 46	40 / 46
2	398 / 198	406 / 182
3	3375 / 1076	3534 / 758
4	27540 / 6534	29208 / 3198
```
Both columns total correctly (2×shared + boundary = 14×tiles), so float32 was failing to merge 8 edge-pairs at iteration 2, 159 at 3, and 1668 at 4 — each one becoming two spurious boundary edges. The fix is verifiably converged, not just better: distinct vertex counts are now identical at rounding 3 through 9 (3734 at iteration 3, 28006 at iteration 4), where float32 gave 5948 at round-6 versus 3841 at round-3.

Side effects: the exact area formula now reproduces to 2e-16 instead of 1.5e-7, and the a/b closure sums to 1.6e-16 instead of 7.5e-8.

What this means for published numbers — mixed_tiling.py is unaffected; I ran it both ways and every figure in that table is identical to four decimals. The braided_tiling "9 tiles, 40 shared, 46 boundary, 120 crossings" line is also unaffected, since it's at 1 iteration. But anything in braid_words_bn.py computed at depth — the depth-7 complexity result in particular — was built on an adjacency graph missing real edges, and should be re-run before you trust it.

+++ spectre.py
@@ -35,7 +35,7 @@ def get_spectre_points(edge_a, edge_b):
 		(        a_d2 - b_sqrt3_d2,     a_sqrt3_d2 + b + b - b_d2), #// 12: -a
 		(0            - b_sqrt3_d2,                  b + b - b_d2), #// 13: -~a
 		(0                        ,                      b       )  #// 14: +~b
-    ], 'float32')
+    ], 'float64')
     # print(spectre_points)
     return spectre_points
    
@@ -43,7 +43,7 @@ SPECTRE_POINTS = get_spectre_points(Edge_a, Edge_b) # tile(Edge_a, Edge_b)
 Mystic_SPECTRE_POINTS = get_spectre_points(Edge_b, Edge_a) # tile(Edge_b, Edge_a)
 SPECTRE_QUAD = SPECTRE_POINTS[[3,5,7,11],:]
 
-IDENTITY = np.array([[1,0,0],[0,1,0]], 'float32') # == trot(0)
+IDENTITY = np.array([[1,0,0],[0,1,0]], 'float64') # == trot(0)
 
 # Rotation matrix for Affine transform
 trot_memo = {
